Retargeting: refactor redirect user composition
