+++ apps/store/src/features/retargeting/RedirectUser.tsx
@@ -1,0 +1,9 @@
+import { RedirectUserParams } from './retargeting.types'
+import { useRedirectUser } from './useRedirectUser'
+
+type Props = RedirectUserParams
+
+export const RedirectUser = (props: Props) => {
+  useRedirectUser(props)
+  return null
+}
