+++ apps/store/src/features/retargeting/retargeting.types.ts
@@ -1,0 +1,4 @@
+export type RedirectUserParams = {
+  shopSessionId: string
+  campaignCode?: string
+}
